reworked create to provide clearer instructions on how to deal with certificate setup
